+++ library/src/schemas/objectWithRest/objectWithRest.ts
@@ -229,7 +229,10 @@ export function objectWithRest(
         // Hint: We exclude specific keys for security reasons
         if (!dataset.issues || !config.abortEarly) {
           for (const key in input) {
-            if (_isValidObjectKey(input, key) && !(key in this.entries)) {
+            if (
+              _isValidObjectKey(input, key) &&
+              !Object.prototype.hasOwnProperty.call(this.entries, key)
+            ) {
               const valueDataset = this.rest['~run'](
                 // @ts-expect-error
                 { value: input[key] },