+++ library/src/schemas/objectWithRest/objectWithRestAsync.ts
@@ -180,7 +180,8 @@ export function objectWithRestAsync(
             Object.entries(input)
               .filter(
                 ([key]) =>
-                  _isValidObjectKey(input, key) && !(key in this.entries)
+                  _isValidObjectKey(input, key) &&
+                  !Object.prototype.hasOwnProperty.call(this.entries, key)
               )
               .map(
                 async ([key, value]) =>